+++ utils/response.go
@@ -17,14 +17,14 @@ type FailedResponse struct {
 	Errors  any    `json:"errors"`
 }
 
-func WriteResponse(c *fiber.Ctx, statusCode int, data interface{}, message string, meta *model.PaginatedMeta) error {
+func WriteResponse(c *fiber.Ctx, statusCode int64, data interface{}, message string, meta *model.PaginatedMeta) error {
 	if data != nil {
 		if errors, ok := data.([]ValidationError); ok {
 			resp := FailedResponse{
 				Message: message,
 				Errors:  errors,
 			}
-			return c.Status(statusCode).JSON(resp)
+			return c.Status(int(statusCode)).JSON(resp)
 		}
 	}
 
@@ -34,10 +34,10 @@ func WriteResponse(c *fiber.Ctx, statusCode int, data interface{}, message strin
 		Meta:    meta,
 	}
 
-	return c.Status(statusCode).JSON(resp)
+	return c.Status(int(statusCode)).JSON(resp)
 }
 
-func WriteErrorResponse(c *fiber.Ctx, statusCode int, message string) error {
+func WriteErrorResponse(c *fiber.Ctx, statusCode int64, message string) error {
 	return WriteResponse(c, statusCode, nil, message, nil)
 }
